+++ lib/src/parser.dart
@@ -64,7 +64,7 @@ class FlatConfig {
     FlatParseOptions options = const FlatParseOptions(),
     LineSplitter lineSplitter = const LineSplitter(),
   }) {
-    assert(!options.commentPrefix.contains(Constants.newline));
+    checkCommentPrefix(options.commentPrefix);
 
     if (source.trim().isEmpty) {
       return FlatDocument.empty();
@@ -93,7 +93,7 @@ class FlatConfig {
     List<String> lines, {
     FlatParseOptions options = const FlatParseOptions(),
   }) {
-    assert(!options.commentPrefix.contains(Constants.newline));
+    checkCommentPrefix(options.commentPrefix);
 
     final out = <FlatEntry>[];
     var lineNumber = 0;
@@ -141,7 +141,7 @@ class FlatConfig {
     Stream<String> stream, {
     FlatParseOptions options = const FlatParseOptions(),
   }) async {
-    assert(!options.commentPrefix.contains(Constants.newline));
+    checkCommentPrefix(options.commentPrefix);
 
     final out = <FlatEntry>[];
 
@@ -193,7 +193,7 @@ class FlatConfig {
     Stream<String> stream, {
     FlatParseOptions options = const FlatParseOptions(),
   }) async* {
-    assert(!options.commentPrefix.contains(Constants.newline));
+    checkCommentPrefix(options.commentPrefix);
 
     var lineNumber = 0;
     await for (var raw in stream) {